+++ et-driver/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-14288] Adding support for DM get asic voltage and set module voltage commands in loopback driver.
 ### Changed
 - [SW-14061] Updating the device mgmt api command codes and Get Module Voltage response name.
 ### Deprecated